Change SelectCommand
SelectCommand now works on tracked parameters instead of workouts
