+++ src/main/java/seedu/address/logic/commands/SelectCommand.java
@@ -11,7 +11,7 @@ import seedu.address.commons.events.ui.JumpToListRequestEvent;
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.workout.Workout;
+import seedu.address.model.workout.Parameter;
 
 /**
  * Selects a person identified using its displayed index from the workout book.
@@ -21,11 +21,11 @@ public class SelectCommand extends Command {
     public static final String COMMAND_WORD = "select";
 
     public static final String MESSAGE_USAGE = COMMAND_WORD
-            + ": Selects the workout identified by the index number used in the displayed workout list.\n"
+            + ": Selects the parameter identified by the index number used in the displayed tracked parameter list.\n"
             + "Parameters: INDEX (must be a positive integer)\n"
             + "Example: " + COMMAND_WORD + " 1";
 
-    public static final String MESSAGE_SELECT_WORKOUT_SUCCESS = "Selected Workout: %1$s";
+    public static final String MESSAGE_SELECT_PARAMETER_SUCCESS = "Selected Parameter: %1$s";
 
 
     private final Index targetIndex;
@@ -38,14 +38,14 @@ public class SelectCommand extends Command {
     public CommandResult execute(Model model, CommandHistory history) throws CommandException {
         requireNonNull(model);
 
-        List<Workout> filteredWorkoutList = model.getFilteredWorkoutList();
+        List<Parameter> filteredTrackedData = model.getFilteredTrackedDataList();
 
-        if (targetIndex.getZeroBased() >= filteredWorkoutList.size()) {
-            throw new CommandException(Messages.MESSAGE_INVALID_WORKOUT_DISPLAYED_INDEX);
+        if (targetIndex.getZeroBased() >= filteredTrackedData.size()) {
+            throw new CommandException(Messages.MESSAGE_INVALID_PARAMETER_DISPLAYED_INDEX);
         }
 
         EventsCenter.getInstance().post(new JumpToListRequestEvent(targetIndex));
-        return new CommandResult(String.format(MESSAGE_SELECT_WORKOUT_SUCCESS, targetIndex.getOneBased()));
+        return new CommandResult(String.format(MESSAGE_SELECT_PARAMETER_SUCCESS, targetIndex.getOneBased()));
 
     }
