+++ src/main/java/seedu/address/ui/TrackedDataListPanel.java
@@ -2,12 +2,16 @@ package seedu.address.ui;
 
 import java.util.logging.Logger;
 
+import com.google.common.eventbus.Subscribe;
+
+import javafx.application.Platform;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.scene.control.ListCell;
 import javafx.scene.control.ListView;
 import javafx.scene.layout.Region;
 import seedu.address.commons.core.LogsCenter;
+import seedu.address.commons.events.ui.JumpToListRequestEvent;
 import seedu.address.commons.events.ui.TrackedDataPanelSelectionChangedEvent;
 import seedu.address.model.workout.Parameter;
 
@@ -41,6 +45,22 @@ public class TrackedDataListPanel extends UiPart<Region> {
                         raise(new TrackedDataPanelSelectionChangedEvent(newValue));
                     }
                 });
+    }
+
+    /**
+     * Scrolls to the {@code TrackedDataListCard} at the {@code index} and selects it.
+     */
+    private void scrollTo(int index) {
+        Platform.runLater(() -> {
+            trackedDataListView.scrollTo(index);
+            trackedDataListView.getSelectionModel().clearAndSelect(index);
+        });
+    }
+
+    @Subscribe
+    private void handleJumpToListRequestEvent(JumpToListRequestEvent event) {
+        logger.info(LogsCenter.getEventHandlingLogMessage(event));
+        scrollTo(event.targetIndex);
     }
 
     /**
